requirements.txt
appnope==0.1.2
asgiref==3.3.4
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.9
dj-database-url==0.5.0
Django==3.2.3
django-shopify-auth==1.2.2
djangorestframework==3.12.4
ecdsa==0.14.1
gunicorn==20.1.0
idna==2.10
install==1.3.4
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyactiveresource==2.2.2
pyasn1==0.4.8
Pygments==2.9.0
PyJWT==2.1.0
python-jose==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
ShopifyAPI==10.0.0
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.4
wcwidth==0.2.5
celery==5.0.5
redis==3.5.3
